split ast
move generate_zcode to codegen.rs
new file evaluate_expression.rs (with all the expression-evaluating stuff)
passage vector for better analyzing (link array in zcode) and nicer error-messages in the future
