+++ src/zwreec/frontend/mod.rs
@@ -25,8 +25,10 @@
 //!     p.parse(tokens)
 //! );
 //! ```
-pub mod lexer;
-pub mod parser;
+
 pub mod ast;
 pub mod codegen;
+pub mod evaluate_expression;
 pub mod expressionparser;
+pub mod lexer;
+pub mod parser;
